// frontend/src/api/models/SpecialtyRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Clinic Management API
 * API documentation for the Clinic Management system
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SpecialtyRequest
 */
export interface SpecialtyRequest {
    /**
     * 
     * @type {string}
     * @memberof SpecialtyRequest
     */
    departmentId?: string;
    /**
     * 
     * @type {string}
     * @memberof SpecialtyRequest
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof SpecialtyRequest
     */
    description?: string;
}

/**
 * Check if a given object implements the SpecialtyRequest interface.
 */
export function instanceOfSpecialtyRequest(value: object): value is SpecialtyRequest {
    return true;
}

export function SpecialtyRequestFromJSON(json: any): SpecialtyRequest {
    return SpecialtyRequestFromJSONTyped(json, false);
}

export function SpecialtyRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): SpecialtyRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'departmentId': json['departmentId'] == null ? undefined : json['departmentId'],
        'name': json['name'] == null ? undefined : json['name'],
        'description': json['description'] == null ? undefined : json['description'],
    };
}

export function SpecialtyRequestToJSON(json: any): SpecialtyRequest {
    return SpecialtyRequestToJSONTyped(json, false);
}

export function SpecialtyRequestToJSONTyped(value?: SpecialtyRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'departmentId': value['departmentId'],
        'name': value['name'],
        'description': value['description'],
    };
}